Refactor: Organize Unity folders, add final updates and comments.
