+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Materials/PointcloudMaterial.mat
