+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Materials/pointClouds.mat
@@ -9,7 +9,7 @@ Material:
   m_PrefabAsset: {fileID: 0}
   m_Name: pointClouds
   m_Shader: {fileID: 4800000, guid: 6a6e7eb2f69ed1043b536e3344061c34, type: 3}
-  m_ShaderKeywords: _ALPHAPREMULTIPLY_ON _GLOSSYREFLECTIONS_OFF _SPECULARHIGHLIGHTS_OFF
+  m_ShaderKeywords: USE_DISTANCE _ALPHAPREMULTIPLY_ON _GLOSSYREFLECTIONS_OFF _SPECULARHIGHLIGHTS_OFF
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
@@ -73,14 +73,15 @@ Material:
     - _Mode: 3
     - _OcclusionStrength: 1
     - _Parallax: 0.02
-    - _PointSize: 1
+    - _PointSize: 10
     - _SmoothnessTextureChannel: 0
     - _SpecularHighlights: 0
     - _SrcBlend: 1
     - _UVSec: 0
+    - _UseDistance: 1
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 1, g: 1, b: 1, a: 1}
+    - _Color: {r: 0.3336609, g: 1, b: 0.13836467, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _ReferencePoint: {r: 280, g: -300, b: 0.7, a: 1}
   m_BuildTextureStacks: []
