--- Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/OffsetVisualizer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47764dd939a52f949b47f3454c9623af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
