+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0aa99b615e69a147b37e74fb9566d01
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
