+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Scripts/DistanceTrigger.cs
@@ -11,23 +11,21 @@ namespace ServerConnection.RosTcpConnector.MANSURI
     public class DistanceTrigger : MonoBehaviour
     {
         public GameObject canvas;
-        public float distanceThreshold = 220f; // Set threshold to 149
         public GameObject BackSphere;
-
-        // The distance that will be set from the STREAMINROS class
-        private float distance = 150f; // Initialize to a value greater than distanceThreshold
-
+        public float distanceThreshold = 220f;
+        //Initialize to a value greater than distanceThreshold
+        private float distance = 280f;
+        // ROS topic from which the distance Data will be subscribed.
         [SerializeField] public string rosDistanceTopic = "/obstacle_distance";
 
-        private Color originalColor; // To store the original color
+        private Color originalColor;
 
-        // Start is called before the first frame update
         void Start()
         {
             // Subscribe to the ROS topic
             ROSConnection.instance.Subscribe<Float64Msg>(rosDistanceTopic, HandleDistanceMessage);
-            Debug.Log("Subscribed to distance topic: " + rosDistanceTopic);
-            Debug.Log(rosDistanceTopic);
+            // Debug.Log("Subscribed to distance topic: " + rosDistanceTopic);
+            // Debug.Log(rosDistanceTopic);
 
             // Store the original color
             RawImage rawImageComponent = BackSphere.GetComponent<RawImage>();
@@ -42,13 +40,13 @@ namespace ServerConnection.RosTcpConnector.MANSURI
         {
             // Update the distance value with the received data
             distance = (float)data.data;
-            Debug.Log("Received data type: " + data.GetType().ToString());
-            Debug.Log("Received distance from /obstacle_distance: " + distance);
+            // Debug.Log("Received data type: " + data.GetType().ToString());
+            // Debug.Log("Received distance from /obstacle_distance: " + distance);
         }
-        
+
         void Update()
         {
-            Debug.Log("Current distance value: " + distance);
+            //Debug.Log("Current distance value: " + distance);
 
             // If the distance is less than or equal to the threshold, show the canvas and change color to red
             if (distance <= distanceThreshold)
@@ -67,7 +65,7 @@ namespace ServerConnection.RosTcpConnector.MANSURI
             // If the distance is greater than the threshold, hide the canvas and revert to original color
             else
             {
-                Debug.Log("Reverting to original color.");
+                // Debug.Log("Reverting to original color.");
 
                 canvas.SetActive(false);
                 RawImage rawImageComponent = BackSphere.GetComponent<RawImage>();
@@ -76,7 +74,7 @@ namespace ServerConnection.RosTcpConnector.MANSURI
                     Material mat = rawImageComponent.material;
 
                     // Convert hex color to Unity's Color format
-                    Color originalColor = new Color(0x5C / 255f, 0x66 / 255f, 0x6E / 255f, 1f); 
+                    Color originalColor = new Color(0x5C / 255f, 0x66 / 255f, 0x6E / 255f, 1f);
 
                     mat.SetColor("_ColorFrom", originalColor);
                     rawImageComponent.material = mat;
