+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Shaders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb47898c999b1e94fa9f5752430391d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
