+++ Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/Shaders/PointClouds2.shader.meta
