--- Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/enableOpenGL.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68eac356a70bf7a449fb80391f40e450
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
