--- Assets/Scripts/ServerConnection/RosTcpConnector/MANSURI/test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47e05b3f5eb5c9045a08a93f8173228d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
