+++ Assets/XR/XRGeneralSettings.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: Android Settings
   m_EditorClassIdentifier: 
   m_LoaderManagerInstance: {fileID: -6955782770196970671}
-  m_InitManagerOnStart: 0
+  m_InitManagerOnStart: 1
 --- !u!114 &-7533418645130189681
 MonoBehaviour:
   m_ObjectHideFlags: 0
